Refactoring the home page
