+++ webapp/src/components/home/FeaturedProducts.tsx
@@ -10,7 +10,7 @@ import Stack from "@mui/material/Stack";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 
-import { checkImageExists } from "../helpers/ImageHelper";
+import { checkImageExists } from "../../helpers/ImageHelper";
 
 const tiers = [
   {
